backend/requirements.txt
Django==5.0.1
whitenoise==6.4.0
wfastcgi==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
social-auth-app-django==5.4.3
django-auth-adfs==1.15.0
colorama==0.4.6
drf_spectacular
pillow
PyLaTex==1.4.1
python-magic==0.4.27

